feat: use ethereum-cryptography for enr crypto
By default, use ethereum-cryptography. Provide a more performant option
for nodejs in the discv5 package, see the discv5 README for details.
