+++ packages/discv5/package.json
@@ -10,6 +10,9 @@
     ".": {
       "import": "./lib/index.js"
     },
+    "./enr/setNodeV4Crypto": {
+      "import": "./lib/enr/setV4Crypto.js"
+    },
     "./kademlia": {
       "import": "./lib/kademlia/index.js"
     },
