+++ packages/enr/src/index.ts
@@ -1,4 +1,6 @@
 export * from "./constants.js";
+export * from "./crypto.js";
+export * as defaultCrypto from "./defaultCrypto.js";
 export * from "./enr.js";
 export * from "./types.js";
 export * from "./create.js";
